+++ tests/index.test-d.ts
@@ -59,20 +59,22 @@ test("any type", () => {
 test("flash", () => {
   const { getSession } = createCookieSessionStorage();
   const session = getSession({} as AstroCookies);
-  expectTypeOf(session.flash.get("success")).toEqualTypeOf<string | undefined>();
+  expectTypeOf(session.flash.get("success")).toEqualTypeOf<
+    string | undefined
+  >();
   expectTypeOf(session.flash.get("notice")).toEqualTypeOf<string | undefined>();
   expectTypeOf(session.flash.get("alert")).toEqualTypeOf<string | undefined>();
   expectTypeOf(session.flash.get("error")).toEqualTypeOf<string | undefined>();
   // @ts-expect-error
   session.flash.set("notice", 1);
   // @ts-expect-error
-  session.flash.get("foo")
+  session.flash.get("foo");
 
   expectTypeOf(session.flash["success"]).toEqualTypeOf<string | undefined>();
   // @ts-expect-error
   session.flash["notice"] = 1;
   // @ts-expect-error
-  session.flash["foo"]
+  session.flash["foo"];
 });
 
 test("flash with type", () => {
@@ -81,9 +83,11 @@ test("flash with type", () => {
   };
   const { getSession } = createCookieSessionStorage<{}, FlashData>();
   const session = getSession({} as AstroCookies);
-  expectTypeOf(session.flash.get("stringKey")).toEqualTypeOf<string | undefined>();
+  expectTypeOf(session.flash.get("stringKey")).toEqualTypeOf<
+    string | undefined
+  >();
   // @ts-expect-error
   session.flash.set("stringKey", 1);
   // @ts-expect-error
-  session.flash.get("foo")
-})
+  session.flash.get("foo");
+});
